fix: preserve long PSS metrics in voice cleanup

+++ app/src/main/java/com/aistudio/mediatool/core/ml/VoiceCleanupProcessor.kt
@@ -319,7 +319,7 @@ class VoiceCleanupProcessor(
         val segmentCount: Int,
         val inferenceMs: Long,
         val realTimeFactor: Double,
-        val peakPssKb: Int,
+        val peakPssKb: Long,
     )
 
     private companion object {
